updated my interface file
