+++ web/src/app/(api)/urls/route.ts
@@ -1,4 +1,3 @@
-import { URLShortenerDto, urlShortenerDtoSchema } from "@/core/schema/url";
 import { connectDB } from "@/server/config/database";
 import { UserDocument } from "@/server/models/user";
 import { urlRepo } from "@/server/repository/url.repo";
@@ -33,18 +32,3 @@ export async function GET() {
     data: urls,
   });
 }
-
-export async function PATCH(req: NextRequest) {
-  await connectDB();
-  const body = (await req.json()) as URLShortenerDto;
-
-  const result = urlShortenerDtoSchema.parse(body);
-  if (!result) {
-    return newBadRequestApiResponse({
-      message: "Validation failed",
-      data: null,
-    });
-  }
-
-  // const updatedUrl = await urlRepo.updatedUrl()
-};
